small scale refactor
